% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hosp.R
\name{aggregate_across_age_groups}
\alias{aggregate_across_age_groups}
\title{Aggregate simulation results into age groups}
\usage{
aggregate_across_age_groups(df, breaks = c(20, 60))
}
\arguments{
\item{df}{simulation output, data frame with columns \code{scenario}, \code{time}, \code{var}, \code{age}, and \code{value}, where \code{age} gives the lower bound of the original five-year age groups}

\item{breaks}{lower bounds for new (aggregated) age groups}
}
\value{
data frame with new aggregated \code{age_group} column
}
\description{
Aggregate simulation results into age groups
}
