% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hosp.R
\name{aggregate_across_epi_subcats}
\alias{aggregate_across_epi_subcats}
\title{Aggregate simulation results into overall epi categories}
\usage{
aggregate_across_epi_subcats(df)
}
\arguments{
\item{df}{simulation output, data frame with columns \code{scenario}, \code{time}, \code{var}, \code{age}, and \code{value}, where \code{age} gives the lower bound of the original five-year age groups}
}
\value{
data frame with values aggregated across epidemiological subcategories
}
\description{
Aggregate simulation results into overall epi categories
}
