% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_values.R
\name{get_default_values}
\alias{get_default_values}
\title{Get named list of default values for a package model}
\usage{
get_default_values(model.name)
}
\arguments{
\item{model.name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}
}
\value{
named list of default values used to initialize model
}
\description{
Get named list of default values for a package model
}
