% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_path.R
\name{get_model_path}
\alias{get_model_path}
\title{Get file path for model files}
\usage{
get_model_path(model.name, file.name, dir.name = "models", local = FALSE)
}
\arguments{
\item{model.name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}

\item{file.name}{(optional) file name; if NULL, return path to model directory}

\item{dir.name}{directory name for files (for local files, must be in main project directory)}

\item{local}{logical. if TRUE, use local model definition (in \code{models} subdirectory of main project repo). if FALSE, use package models.}
}
\value{
a file name complete with file path
}
\description{
For internal use
}
