% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_state}
\alias{get_state}
\title{Get default initial state for a package model}
\usage{
get_state(model_name)
}
\arguments{
\item{model_name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}
}
\value{
Named numeric vector
}
\description{
Get default initial state for a package model
}
