% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Get values for a package model}
\usage{
get_values(model_name, values_file)
}
\arguments{
\item{model_name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}

\item{values_file}{name of the file from which we're loading the values}
}
\value{
values from file
}
\description{
Get values for a package model
}
