% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_flows.R
\name{list_flows}
\alias{list_flows}
\title{List flow names for a model}
\usage{
list_flows(model.name, local = FALSE)
}
\arguments{
\item{model.name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}

\item{local}{logical. if TRUE, use local model definition (in \code{models} subdirectory of main project repo). if FALSE, use package models.}
}
\value{
Character vector of flows in model, see README for \code{model.name} for
explanations of each flow name.
}
\description{
List flow names for a model
}
