% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List available models}
\usage{
list_models(local = FALSE)
}
\arguments{
\item{local}{logical. if TRUE, use local model definition (in \code{models} subdirectory of main project repo). if FALSE, use package models.}
}
\value{
Character vector of model names
}
\description{
These are the names available for use in \code{\link[=get_model]{get_model()}}
}
