% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_five-year-age-groups.R
\name{mk_pop_table}
\alias{mk_pop_table}
\title{Make population table using Mistry et al. data}
\usage{
mk_pop_table(age.group.lower = NULL)
}
\arguments{
\item{age.group.lower}{numeric vector. lower bounds for desired age groups}
}
\value{
a data frame with columns
\itemize{
\item \code{age}: age group
\item \code{count}: population count
}
}
\description{
Make population table using Mistry et al. data
}
