% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hosp.R
\name{plot_output}
\alias{plot_output}
\title{Plot tidied simulation output}
\usage{
plot_output(
  df,
  var_colour = "age",
  var_facet_row = "var",
  var_facet_col = "",
  ylim = NULL,
  facet_scales = "free_y"
)
}
\arguments{
\item{var_colour}{variable name for the colour aesthetic}

\item{var_facet_row}{variable name for facet rows}

\item{var_facet_col}{variable name for facet columns}

\item{ylim}{y-axis limits}

\item{facet_scales}{scales for facets (input into \code{scales} argument of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}})}

\item{output}{tidied simulation output from \code{\link[=tidy_output]{tidy_output()}}}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
Plot tidied simulation output
}
