% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate from a simulator}
\usage{
simulate(simulator, values = NULL)
}
\arguments{
\item{simulator}{Simulator object, initialized using \code{\link[=make_simulator]{make_simulator()}}}

\item{values}{Complete list of model values to update the model simulator before generating the simulation. Must match format of \code{params} list from default (output of \code{\link[=get_default_values]{get_default_values()}} for desired model). If \code{NULL}, use values as currently attached to the model simulator.}
}
\value{
A data frame containing simulation results with columns
\itemize{
\item \code{time}: time in days from simulation start
\item \code{age_group}: age group label of relevant state/flow variable (0 if not an
age structured variable)
\item \code{variable_name}: name of state/flow variable (e.g., "S")
\item \code{value_type}: type of values, such as \code{state} (count of individuals in a compartment) and \code{flow} (flows between compartments)
\item \code{value}
}
}
\description{
Simulate from a simulator
}
