% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_hosp.R
\name{tidy_output}
\alias{tidy_output}
\title{Aggregate simulation results into overall epi categories}
\usage{
tidy_output(output, value_type = "state")
}
\arguments{
\item{output}{simulation output from \code{\link[=simulate]{simulate()}}}

\item{value_type}{value type to keep in tidied output}
}
\value{
data frame with values aggregated across epidemiological subcategories
}
\description{
Aggregate simulation results into overall epi categories
}
