% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Construct a model object from model definition files}
\usage{
get_model(model.name, local = FALSE)
}
\arguments{
\item{model.name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}

\item{local}{logical. if TRUE, use local model definition (in \code{models} subdirectory of main project repo). if FALSE, use package models.}
}
\value{
A \code{\link[macpan2:Model]{macpan2::Model()}} object
}
\description{
Construct a model object from model definition files
}
