% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get a macpan2 model from package models}
\usage{
get_model(model_name)
}
\arguments{
\item{model_name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}
}
\value{
a \code{\link[macpan2:Model]{macpan2::Model()}} object
}
\description{
Get a macpan2 model from package models
}
