% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List models available in the EPACmodel package
(these are the names available for use in \code{\link[=get_model]{get_model()}})}
\usage{
list_models()
}
\value{
character vector of model names
}
\description{
List models available in the EPACmodel package
(these are the names available for use in \code{\link[=get_model]{get_model()}})
}
