% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulator.R
\name{make_simulator}
\alias{make_simulator}
\title{Construct a simulator for a package model}
\usage{
make_simulator(model_name, params = NULL, state = NULL, time_steps = 100)
}
\arguments{
\item{model_name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}

\item{params}{Optional. If NULL, load a default set of parameters}

\item{state}{Optional. If NULL, load a default set of initial states}

\item{time_steps}{Number of time steps for which to run a simulation}
}
\value{
a \code{\link[macpan2:TMBSimulator]{macpan2::TMBSimulator()}} object
}
\description{
Construct a simulator for a package model
}
