% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulator.R
\name{make_simulator}
\alias{make_simulator}
\title{Construct a simulator for a package model}
\usage{
make_simulator(model.name, updated.values = NULL)
}
\arguments{
\item{model.name}{name of the model (see \code{\link[=list_models]{list_models()}} for options)}

\item{updated.values}{Optional. List containing updates to variables + values used to initialize the model simulator. If NULL, use default list is read from disk and used as is.}
}
\value{
a \code{\link[macpan2:TMBSimulator]{macpan2::TMBSimulator()}} object
}
\description{
Construct a simulator for a package model
}
