% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate from a simulator}
\usage{
simulate(simulator)
}
\arguments{
\item{simulator}{Simulator object, initialized using \code{\link[=make_simulator]{make_simulator()}}}
}
\value{
A data frame containing simulation results with columns
\itemize{
\item \code{time}: time in days from simulation start
\item \code{state_name}: compartment name
\item \code{value_type}: type of values, such as \code{state} (count of individuals in a compartment) and \code{total_inflow} (total inflow into a compartment)
\item \code{value}
}
}
\description{
Simulate from a simulator
}
