% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate from a simulator}
\usage{
simulate(simulator)
}
\arguments{
\item{simulator}{Simulator object, initialized using \code{\link[=make_simulator]{make_simulator()}}}
}
\value{
A data frame with the simulation results
}
\description{
Simulate from a simulator
}
